src/lambda/email-retrieval/requirements.txt
boto3==1.34.34
botocore==1.34.34
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
